+++ src/org/example/IAreaCalculable.java
@@ -1,5 +1,5 @@
 package org.example;
 
 public interface IAreaCalculable {
-    float getArea();
+    double getArea();
 }
